// package/entities/Operator.d.ts
export declare class Operator {
    id: string;
    name: string;
    version: string;
    description: string;
    author: string;
    license: string;
    type: string;
    category: string;
    tags?: string;
    codePath: string;
    entryPoint: string;
    operatorType: string;
    inputs?: string;
    outputs?: string;
    operatorParams?: string;
    executionConfig?: string;
    dataVisualization?: string;
    mockdata?: string;
    metadata?: string;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=Operator.d.ts.map